Move Controller to own file
